Extract logging logic from utils
